all: include compile_commands.json in .gitignore
compile_commands.json is used by clangd to get specific information
about the way that the project is compile. it is used to provide more
useful diagnostics and completions.

usually it is created by the build system itself, but Make doesn't
provide it out of the box. however, there are other tools that can
integrate with make and generate a valid compile_commands.json based on
the compilation steps executed by make.

i'm using [bear] to generate compile_commands.json based on the `make
all` command.

[bear]: https://github.com/rizsotto/Bear