tokenizer: remove comment about 'at end'
it seems like this is really how it is defined in the specification.
this makes it really hard to tokenize, since it doensn't fit in with the
rest of the 'name' tokenizers.

maybe it is possible to handle it while parsing, since it only appears
in the :goal section of problems.