build: create static archive target
the reason to build a static library now (before writing the parser)
is because i want to create bindings to pddlp in a higher level
language, such as Rust or Python, and use those bindings to explore
how parsing should work.

i'm not a great C programmer, so being able to express my thoughts in
languages i'm more familiar with should make the process easier.

this does come at the cost of maybe designing an API that is not
easy to implement in C, but it's a risk i'm willing to take :^)